+++ src/cpp/tests/test_abaqus_tools.cpp
@@ -74,7 +74,7 @@ BOOST_AUTO_TEST_CASE( testRowToColumnMajor ){
     BOOST_CHECK( vectorTools::fuzzyEquals( fortran_vector, expected_vector ) );
 }
 
-BOOST_AUTO_TEST_CASE( testExpandAbaqusStandardStressVector ){
+BOOST_AUTO_TEST_CASE( testExpandAbaqusNTENSVector ){
     /*!
      * Test expansion of stress and strain type components to full Abaqus vectors
      */
@@ -90,7 +90,7 @@ BOOST_AUTO_TEST_CASE( testExpandAbaqusStandardStressVector ){
      NDI = 3;
      NSHR = 3;
 
-     vector_expansion = abaqusTools::expandAbaqusStandardStressVector( abaqus_full, NDI, NSHR );
+     vector_expansion = abaqusTools::expandAbaqusNTENSVector( abaqus_full, NDI, NSHR );
 
      BOOST_CHECK( vectorTools::fuzzyEquals( vector_expansion, expected_full ) );
 
@@ -101,12 +101,12 @@ BOOST_AUTO_TEST_CASE( testExpandAbaqusStandardStressVector ){
      NDI = 2;
      NSHR = 1;
 
-     vector_expansion = abaqusTools::expandAbaqusStandardStressVector( abaqus_plane_stress, NDI, NSHR );
+     vector_expansion = abaqusTools::expandAbaqusNTENSVector( abaqus_plane_stress, NDI, NSHR );
 
      BOOST_CHECK( vectorTools::fuzzyEquals( vector_expansion, expected_plane_stress ) );
 }
 
-BOOST_AUTO_TEST_CASE( testContractAbaqusStandardStressVector ){
+BOOST_AUTO_TEST_CASE( testContractAbaqusNTENSVector ){
     /*!
      * Test contraction of stress and strain type components to full Abaqus vectors
      */
@@ -122,7 +122,7 @@ BOOST_AUTO_TEST_CASE( testContractAbaqusStandardStressVector ){
      NDI = 3;
      NSHR = 3;
 
-     vector_contraction_full = abaqusTools::contractAbaqusStandardStressVector( expanded_full, NDI, NSHR );
+     vector_contraction_full = abaqusTools::contractAbaqusNTENSVector( expanded_full, NDI, NSHR );
 
      BOOST_CHECK( vectorTools::fuzzyEquals( vector_contraction_full, abaqus_full ) );
 
@@ -133,13 +133,13 @@ BOOST_AUTO_TEST_CASE( testContractAbaqusStandardStressVector ){
      NDI = 2;
      NSHR = 1;
 
-     vector_contraction_plane_stress = abaqusTools::contractAbaqusStandardStressVector( expanded_plane_stress, NDI, NSHR );
+     vector_contraction_plane_stress = abaqusTools::contractAbaqusNTENSVector( expanded_plane_stress, NDI, NSHR );
 
      BOOST_CHECK( vectorTools::fuzzyEquals( vector_contraction_plane_stress, abaqus_plane_stress ) );
 
 }
 
-BOOST_AUTO_TEST_CASE( testContractAbaqusStandardNTENSMatrix ){
+BOOST_AUTO_TEST_CASE( testContractAbaqusNTENSMatrix ){
     /*!
      * Test contraction of NTENS matrix components from full Abaqus matrices of size 6x6
      */
@@ -165,7 +165,7 @@ BOOST_AUTO_TEST_CASE( testContractAbaqusStandardNTENSMatrix ){
      NDI = 3;
      NSHR = 3;
 
-     matrix_contraction_full = abaqusTools::contractAbaqusStandardNTENSMatrix( expanded_full, NDI, NSHR );
+     matrix_contraction_full = abaqusTools::contractAbaqusNTENSMatrix( expanded_full, NDI, NSHR );
 
      BOOST_CHECK( vectorTools::fuzzyEquals( matrix_contraction_full, abaqus_full ) );
 
@@ -183,7 +183,7 @@ BOOST_AUTO_TEST_CASE( testContractAbaqusStandardNTENSMatrix ){
      NDI = 2;
      NSHR = 1;
 
-     matrix_contraction_plane_stress = abaqusTools::contractAbaqusStandardNTENSMatrix( expanded_plane_stress, NDI, NSHR );
+     matrix_contraction_plane_stress = abaqusTools::contractAbaqusNTENSMatrix( expanded_plane_stress, NDI, NSHR );
 
      BOOST_CHECK( vectorTools::fuzzyEquals( matrix_contraction_plane_stress, abaqus_plane_stress ) );