+++ docs/changelog.rst
@@ -3,8 +3,20 @@
 Changelog
 =========
 
-0.0.2 (unreleased)
+0.0.3 (unreleased)
 ------------------
+
+0.0.2 (2021-04-30)
+------------------
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+- Change Abaqus stress-type vector expansion function name to match the general purpose nature for both Standard and
+  Explicit (:jira:`591`, :pull:`7`). By `Kyle Brindley`_.
+- Change Abaqus stress-type vector contraction function name to match the general purpose nature for both Standard and
+  Explicit (:jira:`592`, :pull:`8`). By `Kyle Brindley`_.
+- Change Abaqus NTENS matrix contraction function name to match the general purpose nature for both Standard and
+  Explicit (:jira:`593`, :pull:`9`). By `Kyle Brindley`_.
 
 0.0.1 (2021-04-13)
 ------------------